refactor(normalizer): rename llm.py -> merchant_normalizer.py
The module is the merchant normalizer, not a generic LLM module (that now
lives at penny/llm.py). Rename for clarity; update the two importers.
