+++ backend/penny/normalizer/__init__.py
@@ -14,7 +14,7 @@ from penny.normalizer.core import (
     slug,
 )
 from penny.normalizer.input import WRAPPER_LABELS, choose_normalizer_input
-from penny.normalizer.llm import MerchantNormalizer
+from penny.normalizer.merchant_normalizer import MerchantNormalizer
 from penny.normalizer.rules import RuleSet, build_system_prompt, load_rules
 
 __all__ = [
